Implement merging of free adjacent blocks when allocating.
Deallocation is now basically free (still requires a write to the given block to
mark it as deallocated, though). When finding a free block, we now merge all
adjacent free blocks we can find on the way, until we find one that is suitably
large. No need for a boundary tag at the end of a block.

Also add 0xDeadBeef to the end of each block :smile: