Extend logging with BeginScope and EndScope.
They will show up as special messages (neither info, nor warning, etc) and increment/decrement the indentation of the given Log.

This implementation is much simpler than ez's LogBlock feature, but it also always prints the begin/end of a block, regardless of whether some regular logging happened within the block. This could be implemented in the future but for now it's not in.